Create Link-Cut-Tree
A wonderful data structure, invented by Robert Tarjan and Daniel Sleator.
For example, we have a task:
Given a graph with n vertices and no edges, number a[i] is written on vertex i 
We have to process 5 type of queries
1) link(u, v) - add an edge between u and v. It's guaranteed, that they were in different connected components.
2) cut(u, v) - remove an edge between u and v. It's guaranteed, that that edge exists
3) connected(u, v) - check whether u and v are connected or not.
4) lca(u, v) - find lca of u and v. It's guaranteed that they're connected
5) sum(v) - find sum of the values on path from v to root of it's connected component
6) sum(u, v) - find sum of the values on path from u to v. It's guaranteed, that they're connected
7) modify(v, x) - make value[v] = x;
8) component_size(v) - find number of nodes in v's connected component

Let's go!
First, how will we store the tree? We will store it's decomposition in implicit BBST with following rules:
1) each vertex has <= 1 children directly connected with it in BBST
2) let's look on node v in BBST. All nodes in it's left subtree must be closer to root than v, in right subtree - farther.
3) each node has 2 pointers(except left and right) - parent and pathparent. Parent is a pointer to splay tree parent, pathparent is a pointer to closest node on the path to root. If node is a root of it's splay tree and root of it's connected component, both parent and pathparent = NULL. 

So, how to process the queries? Without loss of generality, let's suppose than we have initial forest with n vertices and no edges. No one node has parent or pathparent.
The key procedure will be expose(v)
It takes node v and pulls it to up to the root. As the result, root of v's tree and v are in one BBST. 
So, if we've written expose, there is no problem to write link and cut.

Let's think, how to answer LCA query. We can notice, that if we make expose(u) and then expose(v), LCA will be just last node we've visited during expose(v). We are done!
It's also easy to check whether 2 nodes are connected or not - their LCA mustn't be NULL.
How to answer sum queries? Little magic: if we make expose(v), then in field "subtree_sum" in v we will find needed sum on path from v to the root.
Obviously, sum(u, v) can be transformed to sum(u) + sum(v) - 2 * sum(lca(u, v)) + lca(u, v)->value.
To modify value written on vertex, we must expose it and then simply make v->value = x;

To answer component information, we have to maintain "virtual" children information - those children which are not connected with it's parent in BBST.
Read https://codeforces.com/blog/entry/67637 for more information about this trick.

If we use splay tree, we can process queries in O(log n) amortized (because expose runs in O(log n) on average). If we use Cartesian tree or Red-Black tree etc., total time complexity will be O(log^2(n))per query. There will be no proofs.