Create SCC + 2-SAT
2-SAT is a problem where we have n booleans and m logical expressions. Our task is to assign 0 or 1 to each boolean and make all conditions satisfied or say that there's no solution.
So, we make digraph on 2 * n vertices and transform each condition to 2 implication edges (or more). Then we find strongly connected components of this graph(scc is the max. subset of vertices such that for each (u, v) (u ∈ S && v ∈ S) u is reachable from v and v is reachable from u. 
So, let's look at our vertices. Vertex 2 * i corresponding to ith boolean = false, and 2 * i + 1 to ith boolean = true.
If they're in the same scc, obviously, there is no solution, because whichever value we choose, we will come to a contradiction. Else we can easily determine, which value do we have to assign to ith boolean. If scc[2 * i] is less than scc[2 * i + 1], it means that 2 * i + 1 is reachable from 2 * i, but 2 * i isn't reachable from 2 * i + 1, because if they are reachable from each other, they must be in one scc. That's all!
Time complexity - O(n + m), where n = number of booleans, m = number of conditions(because each condition can be expressed with constant number of edges)